abilities now work properly, set up the cooldown shader with the cooldown time for individual abilities, add a speed up ability

+++ Assets/prefabs/UI/InGameUI.cs
@@ -35,7 +35,7 @@ public class InGameUI : MonoBehaviour
         PlayerHealthComp.BroadCastCurrentHealth();
     }
 
-    private void PlayerHealthChanged(int newValue, int oldValue, int maxValue, GameObject Causer)
+    private void PlayerHealthChanged(float newValue, float oldValue, float maxValue, GameObject Causer)
     {
         SetPlayerHealth((float)newValue/(float)maxValue);
     }
